Print every other line from the input...
Another requirement: starting with the first line.

To get all the inputs once (which are lines) we used command paste --
The command outputs lines consisting of lines from each file specified,
separated by tab as delimiter to the standard output

In order to get just the first line (field) we used the command cut -f1.
Option -d is not specified and its default value(TAB) is considered.